+++ sklearn/isotonic.py
@@ -8,7 +8,8 @@ from scipy import interpolate
 from scipy.stats import spearmanr
 from .base import BaseEstimator, TransformerMixin, RegressorMixin
 from .utils import as_float_array, check_array, check_consistent_length
-from ._isotonic import _isotonic_regression
+from .utils.fixes import astype
+from ._isotonic import _isotonic_regression, _make_unique
 import warnings
 import math
 
@@ -200,12 +201,24 @@ class IsotonicRegression(BaseEstimator, TransformerMixin, RegressorMixin):
         The stepwise interpolating function that covers the domain
         X_.
 
+    Notes
+    -----
+    Ties are broken using the secondary method from Leeuw, 1977.
+
     References
     ----------
     Isotonic Median Regression: A Linear Programming Approach
     Nilotpal Chakravarti
     Mathematics of Operations Research
     Vol. 14, No. 2 (May, 1989), pp. 303-308
+
+    Isotone Optimization in R : Pool-Adjacent-Violators
+    Algorithm (PAVA) and Active Set Methods
+    Leeuw, Hornik, Mair
+    Journal of Statistical Software 2009
+
+    Correctness of Kruskal's algorithms for monotone regression with ties
+    Leeuw, Psychometrica, 1977
     """
     def __init__(self, y_min=None, y_max=None, increasing=True,
                  out_of_bounds='nan'):
@@ -228,8 +241,12 @@ class IsotonicRegression(BaseEstimator, TransformerMixin, RegressorMixin):
                              .format(self.out_of_bounds))
 
         bounds_error = self.out_of_bounds == "raise"
-        self.f_ = interpolate.interp1d(X, y, kind='slinear',
-                                       bounds_error=bounds_error)
+        if len(y) == 1:
+            # single y, constant prediction
+            self.f_ = lambda x: y.repeat(x.shape)
+        else:
+            self.f_ = interpolate.interp1d(X, y, kind='slinear',
+                                           bounds_error=bounds_error)
 
     def _build_y(self, X, y, sample_weight):
         """Build the y_ IsotonicRegression."""
@@ -249,8 +266,13 @@ class IsotonicRegression(BaseEstimator, TransformerMixin, RegressorMixin):
 
         order = np.lexsort((y, X))
         order_inv = np.argsort(order)
-        self.X_ = as_float_array(X[order], copy=False)
-        self.y_ = isotonic_regression(y[order], sample_weight, self.y_min,
+        if sample_weight is None:
+            sample_weight = np.ones(len(y))
+        X, y, sample_weight = [astype(array[order], np.float64, copy=False)
+                               for array in [X, y, sample_weight]]
+        unique_X, unique_y, unique_sample_weight = _make_unique(X, y, sample_weight)
+        self.X_ = unique_X
+        self.y_ = isotonic_regression(unique_y, unique_sample_weight, self.y_min,
                                       self.y_max, increasing=self.increasing_)
 
         return order_inv
@@ -318,44 +340,6 @@ class IsotonicRegression(BaseEstimator, TransformerMixin, RegressorMixin):
         if self.out_of_bounds == "clip":
             T = np.clip(T, self.X_min_, self.X_max_)
         return self.f_(T)
-
-    def fit_transform(self, X, y, sample_weight=None):
-        """Fit model and transform y by linear interpolation.
-
-        Parameters
-        ----------
-        X : array-like, shape=(n_samples,)
-            Training data.
-
-        y : array-like, shape=(n_samples,)
-            Training target.
-
-        sample_weight : array-like, shape=(n_samples,), optional, default: None
-            Weights. If set to None, all weights will be equal to 1 (equal
-            weights).
-
-        Returns
-        -------
-        y_ : array, shape=(n_samples,)
-            The transformed data.
-
-        Notes
-        -----
-        X doesn't influence the result of `fit_transform`. It is however stored
-        for future use, as `transform` needs X to interpolate new input
-        data.
-        """
-        # Build y_
-        order_inv = self._build_y(X, y, sample_weight)
-
-        # Handle the left and right bounds on X
-        self.X_min_ = np.min(self.X_)
-        self.X_max_ = np.max(self.X_)
-
-        # Build f_
-        self._build_f(self.X_, self.y_)
-
-        return self.y_[order_inv]
 
     def predict(self, T):
         """Predict new data by linear interpolation.